Created module for Thanos AI. Created shared module for models and DTOs.
