+++ server/SQLPlugin/SQLPlugin.iml
@@ -9,6 +9,6 @@
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="module" module-name="PluginInterfaces" scope="PROVIDED" />
     <orderEntry type="library" name="KotlinJavaRuntime" level="project" />
-    <orderEntry type="library" name="org.xerial:sqlite-jdbc:3.27.2" level="project" />
+    <orderEntry type="library" name="org.xerial:sqlite-jdbc:3.27.2.1" level="project" />
   </component>
 </module>
