+++ server/Thanos-AI/Thanos-AI.iml
@@ -7,6 +7,7 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="KotlinJavaRuntime" level="project" />
+    <orderEntry type="library" name="KotlinJavaRuntime (4)" level="project" />
+    <orderEntry type="module" module-name="Shared Models" />
   </component>
 </module>
