+++ server/Thanos-AI/src/Thanos.kt
@@ -1,0 +1,6 @@
+/**
+ * Created by Jordan Gassaway on 4/22/2019.
+ * Thanos TODO: Enter Class Purpose Here
+ */
+class Thanos {
+}
