+++ server/src/models/Games.kt
@@ -1,7 +1,10 @@
 package models
 
+import Color
+import GamePlayerDTO
 import commands.*
 import IGame
+import LobbyGameDTO
 import persistence.PersistenceManager
 import java.lang.Integer.max
